devops: Add pre-commit and pre-push hooks
These git hooks are installed on a developer's machine through running `composer install` or running the setup script with the `--fresh` option. The main reason I decided to go the path of creating these hooks through composer was because it cuts down on making 2 new bash scripts that effectively run one command and a bash script that copies them into the developer's `.git` directory as `brainmaestro/composer-git-hooks` does all of that heavy lifting after an install of all dependencies. Along with that, multiple resources online were saying that if you are developing a project that is using composer then its the most standard way to transfer pre-commit and pre-push hooks to all developers on the project.