refactor: Change how to determine when to encode the output for `sanitizeStoredFeedback`
Add `is_vulnerable` attribute to a feedback entry so that when the user exploits Sotred XSS on the page when it is vulnerable, it makes a note that it is exploit code. When the user switches the page back to secure, it will only encode the exploit entries as every other entry should be HTML escaped by default.